Let the example show a generated scope, not a typed one
It spelled `ClassScope('Hero', 'lz7xyh')` by hand under a comment saying
the builder writes it — the one thing the package tells you never to do.
Now the example is the real arrangement: `hero.dart` carries the part
directive and `@scopedCss`, `hero.scopes.dart` beside it is what the
builder would write for that asset, down to the suffix, which is the md5
of `jaspr_class_scope|example/hero.dart#Hero`.
