+++ packages/jaspr_class_scope/example/hero.scopes.dart
@@ -1,0 +1,8 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// Written by jaspr_class_scope_builder.
+
+part of 'hero.dart';
+
+/// The class-name scope of [Hero], hashed from
+/// `jaspr_class_scope|example/hero.dart#Hero`.
+const _$heroScope = ClassScope('Hero', '7en7oj');
