+++ packages/jaspr_class_scope/example/jaspr_class_scope_example.dart
@@ -1,32 +1,10 @@
-// The Jaspr parts — `@css`, `css()`, `div()` — are left out so that the
-// example stays dependency-free.
-
 // The example prints what the component would render.
 // ignore_for_file: avoid_print
 
-import 'package:jaspr_class_scope/jaspr_class_scope.dart';
-
-/// What the builder writes into `hero.scopes.dart`.
-const _$heroScope = ClassScope('Hero', 'lz7xyh');
-
-class Hero {
-  static const _class = _$heroScope;
-
-  static final _grid = _class('grid');
-  static final _title = _class('title');
-
-  /// What a `@css` getter would style.
-  static String get styles => '${_grid.selector} > ${_title.selector}';
-
-  /// What `build` would render.
-  static String get gridClasses => _grid.name;
-}
-
-/// A class the page's script looks up by name is not scoped.
-const copyButton = ClassName.shared('js-copy');
+import 'hero.dart';
 
 void main() {
-  print(Hero.styles); // .grid-lz7xyh > .title-lz7xyh
-  print(Hero.gridClasses); // grid-lz7xyh
+  print(Hero.styles); // .grid-7en7oj > .title-7en7oj
+  print(Hero.gridClasses); // grid-7en7oj
   print(copyButton.name); // js-copy
 }
